# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuildSure API
  description: FastAPI backend for BuildSure project management system with complete CRUD operations for organizations, users, projects, and sections.
  version: 2.0.0
  contact:
    name: BuildSure Team
    email: support@buildsure.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.buildsure.com
    description: Production server
...
paths:
...
  /api/v1/projects/:
    get:
      summary: List projects
      description: Get paginated list of projects with optional filtering
      tags:
        - Projects
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: org_id
          in: query
          description: Filter by organization ID
          required: false
          schema:
            type: string
            format: uuid
        - name: user_id
          in: query
          description: Filter by user ID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'

    post:
      summary: Create project
      description: Create a new project
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailsResponse'
        '400':
          description: Bad request (organization or user not found, or user doesn't belong to organization)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_id}:
    get:
      summary: Get project
      description: Get project by ID
      tags:
        - Projects
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailsResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update project
      description: Update project by ID
      tags:
        - Projects
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailsResponse'
        '400':
          description: Bad request (organization or user not found, or user doesn't belong to organization)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete project
      description: Delete project by ID (cascades to sections)
      tags:
        - Projects
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Project deleted
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/organizations/{org_id}/projects:
    get:
      summary: List projects by organization
      description: Get paginated list of projects for a specific organization
      tags:
        - Projects
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: List of projects in organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/users/{user_id}/projects:
    get:
      summary: List projects by user
      description: Get paginated list of projects for a specific user
      tags:
        - Projects
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: List of projects for user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_id}/report:
    get:
      summary: Get project report
      description: Generate and return a comprehensive report for the project in various formats (JSON, PDF, Excel)
      tags:
        - Projects
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          description: Report format
          required: false
          schema:
            type: string
            enum: [json, pdf, excel]
            default: json
      responses:
        '200':
          description: Project report in JSON format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectReportResponse'
            application/pdf:
              schema:
                type: string
                format: binary
              example: "PDF file content"
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
              example: "Excel file content"
        '400':
          description: Invalid format parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to generate report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ProjectStatus:
      type: string
      enum:
        - not_started
        - in_progress
        - completed
        - on_hold
      example: "not_started"
    ProjectBase:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 255
          example: "Website Redesign"
        description:
          type: string
          nullable: true
          example: "Complete redesign of company website"
        status:
          $ref: '#/components/schemas/ProjectStatus'
        current_section:
          type: integer
          minimum: 0
          default: 0
          example: 0
        total_sections:
          type: integer
          minimum: 0
          default: 0
          example: 5
        completed_sections:
          type: integer
          minimum: 0
          default: 0
          example: 0
    ProjectCreate:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'
        - type: object
          required:
            - org_id
            - user_id
          properties:
            org_id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            user_id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174001"
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          example: "Website Redesign"
        description:
          type: string
          nullable: true
          example: "Complete redesign of company website"
        status:
          $ref: '#/components/schemas/ProjectStatus'
        current_section:
          type: integer
          minimum: 0
          example: 1
        total_sections:
          type: integer
          minimum: 0
          example: 5
        completed_sections:
          type: integer
          minimum: 0
          example: 1
        org_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        user_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
    ProjectDetailsResponse:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'
        - type: object
          required:
            - id
            - org_id
            - user_id
            - created_at
            - updated_at
          properties:
            id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174002"
            org_id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            user_id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174001"
            created_at:
              type: string
              format: date-time
              example: "2023-01-01T12:00:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2023-01-01T12:00:00Z"
    ProjectListResponse:
      type: object
      required:
        - items
        - total
        - page
        - size
        - pages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectDetailsResponse'
        total:
          type: integer
          example: 1
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        pages:
          type: integer
          example: 1

    # Section Schemas
...
    SectionReportData:
      type: object
      required:
        - section_number
        - completed
      properties:
        section_number:
          type: integer
          minimum: 1
          example: 1
        final_output:
          type: object
          nullable: true
          example:
            deliverable: "Complete login system"
            files: ["login.js", "auth.js"]
        completed:
          type: boolean
          example: true
    ProjectReportResponse:
      type: object
      required:
        - project_id
        - project_name
        - project_status
        - total_sections
        - completed_sections
        - generated_at
        - sections
      properties:
        project_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174002"
        project_name:
          type: string
          example: "Website Redesign"
        project_status:
          $ref: '#/components/schemas/ProjectStatus'
        total_sections:
          type: integer
          example: 26
        completed_sections:
          type: integer
          example: 15
        generated_at:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
        sections:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SectionReportData'
          example:
            section_1:
              section_number: 1
              final_output:
                deliverable: "Project requirements analysis"
                files: ["requirements.md"]
              completed: true
            section_2:
              section_number: 2
              final_output: null
              completed: false

    # Answer Schemas
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: "Organization not found"